# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SMSOne Enterprise API
  description: |-
    Enterprise Spring Modulith template API. Every response uses the unified envelope ({data | errors, meta, links}) with meta.requestId always present; quote the requestId when reporting issues.

    Operations are grouped as `<axis> · <resource>`. The axis is the authority the operation enforces — Platform (a realm role), Organization (a permission inside one tenant), Shared (any authenticated caller) — and it is read from the code that enforces it, so a group name cannot promise access the API will refuse.
  contact:
    name: SMSOne
    email: ayesigapo@gmail.com
  version: "1"
servers:
- url: http://localhost:8080
  description: Local
- url: https://staging-api.smsone.co.ug
  description: Staging
- url: https://api.smsone.co.ug
  description: Production
security:
- bearerAuth: []
- keycloak: []
...
paths:
...
  /api/v1/orgs/{orgId}/webhooks/{id}:
    get:
      tags:
      - Organization · Webhooks
      summary: Get one webhook subscription
      description: The `secret` attribute is masked here — it is shown in full only
        at create.
      operationId: get_1
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceObject"
    put:
      tags:
      - Organization · Webhooks
      summary: Update a webhook subscription
      description: "`url` and `events` are replaced wholesale, not merged. Omitting\
        \ `status` resets the subscription to ACTIVE, so a disabled subscription silently\
        \ resumes delivery unless the update repeats `status: DISABLED`. The signing\
        \ secret is never rotated here."
      operationId: update
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWebhookRequest"
        required: true
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceObject"
    delete:
      tags:
      - Organization · Webhooks
      summary: Delete a webhook subscription
      description: "Stops matching new events and cancels everything already queued\
        \ for the endpoint, so a revoked URL is not retried for hours. The delivery\
        \ log survives and stays readable — \"what did we send that endpoint?\" is\
        \ exactly the question asked after a delete."
      operationId: delete
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
...
  /api/v1/orgs/{orgId}/webhooks:
    get:
      tags:
      - Organization · Webhooks
      summary: List the organization's webhook subscriptions
      operationId: list
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: "page[size]"
        in: query
        description: "Items per page, 1–100. Defaults to 20."
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - name: "page[after]"
        in: query
        description: Opaque keyset cursor taken from a previous response's meta.page.nextCursor.
          Omit for the first page.
        required: false
        schema:
          type: string
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WindowedResultResourceObject"
    post:
      tags:
      - Organization · Webhooks
      summary: Register a webhook subscription
      description: "The signing secret is returned in full exactly once — in this\
        \ response. Every later read masks it and there is no way to retrieve it again,\
        \ so store it now: it is the key that verifies the `X-Webhook-Signature` HMAC\
        \ on delivered events."
      operationId: create_1
      parameters:
      - name: orgId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Impersonate
        in: header
        description: "Run this request as the target of an impersonation session (its\
          \ id, from POST /api/v1/admin/impersonations). The session must belong to\
          \ the calling operator; a READ_ONLY session refuses any unsafe method. Every\
          \ write is audited against the operator, not the target."
        required: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
        required: true
      responses:
        "201":
          description: Created
          headers:
            X-Request-Id:
              description: "Public request id (accepted inbound, else minted as a\
                \ ULID)."
              schema:
                type: string
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResourceObject"
...
components:
  schemas:
...
    ResourceObject:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        attributes: {}
    UpdateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          minLength: 1
        events:
          type: array
          items:
            type: string
          minItems: 1
          uniqueItems: true
        status:
          type: string
      required:
      - events
      - url
...
    CreateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          minLength: 1
        events:
          type: array
          items:
            type: string
          minItems: 1
          uniqueItems: true
      required:
      - events
      - url
...
    PageMeta:
      type: object
      properties:
        size:
          type: integer
          format: int32
        count:
          type: integer
          format: int32
        hasMore:
          type: boolean
        nextCursor:
          type: string
    WindowedResultResourceObject:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResourceObject"
        page:
          $ref: "#/components/schemas/PageMeta"
  securitySchemes:
    bearerAuth:
      type: http
      description: Paste a Keycloak-issued access token.
      scheme: bearer
      bearerFormat: JWT
    keycloak:
      type: oauth2
      description: Run the Keycloak authorization-code flow (PKCE).
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8081/realms/smsone/protocol/openid-connect/auth
          tokenUrl: http://localhost:8081/realms/smsone/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect
            profile: Profile claims